Move getDefaultFigureForType to locked-figures dir
