+++ packages/perseus-editor/src/components/__stories__/color-select.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import ColorSelect from "../../widgets/interactive-graph-editor/locked-figures/color-select";
-import {getDefaultFigureForType} from "../util";
+import {getDefaultFigureForType} from "../../widgets/interactive-graph-editor/locked-figures/util";
 
 import type {LockedFigureColor} from "@khanacademy/perseus";
 import type {Meta} from "@storybook/react";
