+++ packages/perseus-editor/src/components/__stories__/color-swatch.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import ColorSwatch from "../../widgets/interactive-graph-editor/locked-figures/color-swatch";
-import {getDefaultFigureForType} from "../util";
+import {getDefaultFigureForType} from "../../widgets/interactive-graph-editor/locked-figures/util";
 
 import type {Meta} from "@storybook/react";
