+++ packages/perseus-editor/src/components/__stories__/locked-function-settings.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import LockedFunctionSettings from "../../widgets/interactive-graph-editor/locked-figures/locked-function-settings";
-import {getDefaultFigureForType} from "../util";
+import {getDefaultFigureForType} from "../../widgets/interactive-graph-editor/locked-figures/util";
 
 import type {Meta, StoryObj} from "@storybook/react";
