+++ packages/perseus-editor/src/components/__stories__/locked-point-settings.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import LockedPointSettings from "../../widgets/interactive-graph-editor/locked-figures/locked-point-settings";
-import {getDefaultFigureForType} from "../util";
+import {getDefaultFigureForType} from "../../widgets/interactive-graph-editor/locked-figures/util";
 
 import type {Meta, StoryObj} from "@storybook/react";
