+++ packages/perseus-editor/src/components/__stories__/locked-polygon-settings.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import LockedPolygonSettings from "../../widgets/interactive-graph-editor/locked-figures/locked-polygon-settings";
-import {getDefaultFigureForType} from "../util";
+import {getDefaultFigureForType} from "../../widgets/interactive-graph-editor/locked-figures/util";
 
 import type {Meta, StoryObj} from "@storybook/react";
