+++ packages/perseus-editor/src/components/__tests__/util.test.ts
@@ -1,6 +1,5 @@
 import {
     degreeToRadian,
-    getDefaultFigureForType,
     radianToDegree,
     getDefaultGraphStartCoords,
     getSinusoidEquation,
@@ -9,107 +8,6 @@ import {
 } from "../util";
 
 import type {PerseusGraphType, Range} from "@khanacademy/perseus";
-
-describe("getDefaultFigureForType", () => {
-    test("should return a point with default values", () => {
-        const figure = getDefaultFigureForType("point");
-        expect(figure).toEqual({
-            type: "point",
-            coord: [0, 0],
-            color: "grayH",
-            filled: true,
-        });
-    });
-
-    test("should return a line with default values", () => {
-        const figure = getDefaultFigureForType("line");
-        expect(figure).toEqual({
-            type: "line",
-            kind: "line",
-            points: [
-                {
-                    type: "point",
-                    coord: [0, 0],
-                    color: "grayH",
-                    filled: true,
-                },
-                {
-                    type: "point",
-                    coord: [2, 2],
-                    color: "grayH",
-                    filled: true,
-                },
-            ],
-            color: "grayH",
-            lineStyle: "solid",
-            showPoint1: false,
-            showPoint2: false,
-        });
-    });
-
-    test("should return a vector with default values", () => {
-        const figure = getDefaultFigureForType("vector");
-        expect(figure).toEqual({
-            type: "vector",
-            points: [
-                [0, 0],
-                [2, 2],
-            ],
-            color: "grayH",
-        });
-    });
-
-    test("should return an ellipse with default values", () => {
-        const figure = getDefaultFigureForType("ellipse");
-        expect(figure).toEqual({
-            type: "ellipse",
-            center: [0, 0],
-            radius: [1, 1],
-            angle: 0,
-            color: "grayH",
-            fillStyle: "none",
-            strokeStyle: "solid",
-        });
-    });
-
-    test("should return a polygon with default values", () => {
-        const figure = getDefaultFigureForType("polygon");
-        expect(figure).toEqual({
-            type: "polygon",
-            points: [
-                [0, 2],
-                [-1, 0],
-                [1, 0],
-            ],
-            color: "grayH",
-            showVertices: false,
-            fillStyle: "none",
-            strokeStyle: "solid",
-        });
-    });
-
-    test("should return a 'function' with default values", () => {
-        const figure = getDefaultFigureForType("function");
-        expect(figure).toEqual({
-            type: "function",
-            color: "grayH",
-            strokeStyle: "solid",
-            equation: "x^2",
-            directionalAxis: "x",
-        });
-    });
-
-    test("should return a 'label' with default values", () => {
-        const figure = getDefaultFigureForType("label");
-        expect(figure).toEqual({
-            type: "label",
-            coord: [0, 0],
-            text: "",
-            color: "grayH",
-            size: "medium",
-        });
-    });
-});
 
 describe("degreeToRadian", () => {
     test.each`
