+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/locked-figures-section.test.tsx
@@ -4,7 +4,7 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
 import LockedFiguresSection from "./locked-figures-section";
-import {getDefaultFigureForType} from "../../../components/util";
+import {getDefaultFigureForType} from "./util";
 
 import type {UserEvent} from "@testing-library/user-event";
